\name{bws3Interactions}
\alias{bws3Interactions}

\title{Create Interactions between Attributes/Levels and Covariates}

\description{
Create interactions between attribute/level variables and covariates in
the data set for BWS3 analysis.
}

\details{
This dialog box creates interactions between attribute/level variables and
covariates in the data set for BWS3 analysis. 
This dialog box is activated if the data set for BWS3 analysis created in 
the \sQuote{\code{Create Data Set for BWS3 Analysis}} dialog box is selected 
as the active data set.

GUI elements of this dialog box are as follows.

\itemize{
\item \code{Attribute/level variables}: Variables to be used as 
attribute and/or level variables in the model. Select one or more 
variables from this list box.

\item \code{Covariates}: Variables to be used as covariates in the model. 
Select one or more variables from this list box.
}
}

