\name{bws3Questions}
\alias{bws3Questions}

\title{Display BWS3 Questions}

\description{
Display a series of questions for BWS3.
}

\details{
This dialog box shows a series of BWS3 questions that are converted from
the choice sets designed in the \sQuote{\code{Design Choice Sets for BWS3}}
dialog box using the function \code{\link[support.CEs]{questionnaire}} in
\pkg{support.CEs}. 
For details on the results of this dialog box, refer to help for 
\code{\link[support.CEs]{questionnaire}}. 

GUI element of this dialog box is as follows.

\itemize{
\item \code{Design}: An R object name containing BWS3 choice sets generated
in the \sQuote{\code{Design Choice Sets for BWS3}} dialog box. 
The default name \sQuote{\code{BWS3design}} is displayed automatically. 
The correct name should be entered if the name of BWS3 choice sets 
(\code{BWS3design}) was changed in the dialog box for BWS3 design.
}
}

